+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add basic session example
 - Configure CDN to forward cookies
 - Expose RemixStack CDK components
+- Update deployed code to use `jwtCookieSessionHandlerFactory`
 
 ## 2020-11-29